Move test verification code dialog to Dialogs class
