+++ app/src/main/res/values/strings.xml
@@ -247,5 +247,6 @@
 
     <string name="dialog_test_verification_code_expl_title">What are they?</string>
     <string name="dialog_test_verification_code_expl_text">They are strings of numbers and letters that are provided by The Public Health Department to accurately identify valid positive COVID-19 diagnoses. They help protect users of the app by preventing the entry of inacurrate or fraudulent test results.</string>
+    <string name="dialog_test_verification_code_expl_more">Learn more</string>
 
 </resources>
